Add files via upload
This is my first project using python and tkinter and making an app and My App name is Flashcard Based Language Learning System
The Project highlights not only the power of python programming but also the significance of creating software that addresses real-world learning needs. By offering a reliable and intuitive platform for language acquisition, the Flashcard- Based Language Learning System exemplifies the potential of educational technology in making language learning accessible, effective, and enjoyable.

The main objectives of this are given below:

To develop a user-friendly application for learning languages through flashcards.
To provide an interactive language learning experience with flashcard.
To implement progressive learning levels with a focus on mastery.
To Enhance language learning with pronunciation hints and gamified progression.
To Ensure scalability for future enhance and support for additional language learning tools.

The Scope of the project
To facilitate language learning by offering an interactive, user-friendly interface that allow users to learn new languages through flashcards. 
Supports multiple languages, including Hindi, Spanish, French, Telugu, and German, with categories such as Alphabets, Numbers, Basic words and Greetings. The application features a progressive learning system with three levels of difficulty -Easy, Medium and Hard ensuring that user master each level before progressing.
 In Conclusion, the modular code structure ensures that the application can be expanded to support more languages, advanced tools, and interactive elements, making it a versatile platform for language learners.
Here the user gets the warm welcome before starting a great journey and also there are two options provided 
Start Learning Now
Explore More
You can learn languages using flashcards by clicking on Start Learning Now and If you click on Explore more you will get a pop-up message telling that Stay tune for more features.
This is for the user who are back of Gen-z because they will like this sort of interaction
In this page the user  will be able to choose the preferred language and what he wants to learn particularly.  And later  just below the selected option user will be able to learn a new language through flashcards.
The user can learn anything and then explore what they want to learn and tan also learn many languages simultaneously.
In the previous slide  we saw a picture showing Number One as Uno in the language Spanish and how they present the number in English is this way and they was shown when the user clicked on FLIP option and the user can go to the next number if the user click on the option NEXT. Can also go through the previous number by clicking in the PREVIOUS option. 
It’s easy for all the lazy learners around.
When the user chose this option then a pop out comes telling 
“Stay Tuned! More features are coming soon. Follow us for updates!!